Implement interfaces for keydb storage
