+++ common/basecomponent/base.go
@@ -138,7 +138,7 @@ func (b *BaseDendrite) CreateAccountsDB() *accounts.Database {
 
 // CreateKeyDB creates a new instance of the key database. Should only be called
 // once per component.
-func (b *BaseDendrite) CreateKeyDB() *keydb.Database {
+func (b *BaseDendrite) CreateKeyDB() keydb.Database {
 	db, err := keydb.NewDatabase(string(b.Cfg.Database.ServerKey))
 	if err != nil {
 		logrus.WithError(err).Panicf("failed to connect to keys db")
