+++ common/keydb/postgres/server_key_table.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package keydb
+package postgres
 
 import (
 	"context"
